rename notes panel + other changes to match comment requests
