+++ src/components/settings/SettingsHeader.tsx
@@ -9,7 +9,7 @@ type SettingsHeaderProps = {
 
 export default function SettingsHeader({ user }: SettingsHeaderProps) {
   return (
-    <Panel className="relative bg-[linear-gradient(106deg,#C0C2FF_13.64%,#DDBBF3_48.08%,#FFC6C1_83.43%)]">
+    <Panel className={`relative ${HOME_PAGE_GRADIENT_CLASS}`}>
       <div className="absolute inset-0 dark:bg-slightly-darken" />
       <div className="z-10">
         <div className="flex gap-4 max-sm:flex-col-reverse sm:flex-row pl-2">
